Move ProcessRun to core module
