+++ java/clients/src/main/java/sleeper/status/report/compactionjob/StandardCompactionJobStatusReporter.java
@@ -17,8 +17,8 @@
 package sleeper.status.report.compactionjob;
 
 import sleeper.compaction.job.status.CompactionJobStatus;
-import sleeper.compaction.job.status.ProcessRun;
 import sleeper.core.record.process.AverageRecordRate;
+import sleeper.core.status.ProcessRun;
 import sleeper.status.report.table.TableField;
 import sleeper.status.report.table.TableRow;
 import sleeper.status.report.table.TableWriter;
