+++ java/compaction-core/src/main/java/sleeper/compaction/job/status/CompactionJobStatusesBuilder.java
@@ -16,6 +16,7 @@
 package sleeper.compaction.job.status;
 
 import sleeper.core.status.ProcessFinishedStatus;
+import sleeper.core.status.ProcessRun;
 import sleeper.core.status.ProcessStartedStatus;
 import sleeper.core.status.ProcessStatusUpdate;
