+++ java/compaction-core/src/test/java/sleeper/compaction/job/CompactionJobRunTest.java
@@ -19,10 +19,10 @@ package sleeper.compaction.job;
 import org.junit.Test;
 import sleeper.compaction.job.status.CompactionJobCreatedStatus;
 import sleeper.compaction.job.status.CompactionJobStatus;
-import sleeper.compaction.job.status.ProcessRun;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.status.ProcessFinishedStatus;
+import sleeper.core.status.ProcessRun;
 import sleeper.core.status.ProcessStartedStatus;
 
 import java.time.Instant;
@@ -30,7 +30,7 @@ import java.time.Instant;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.groups.Tuple.tuple;
 
-public class ProcessRunTest {
+public class CompactionJobRunTest {
     private static final String DEFAULT_TASK_ID_1 = "task-id-1";
     private static final String DEFAULT_TASK_ID_2 = "task-id-2";
