+++ java/compaction-core/src/test/java/sleeper/compaction/job/CompactionJobStatusInPeriodTest.java
@@ -18,10 +18,10 @@ package sleeper.compaction.job;
 import org.junit.Test;
 import sleeper.compaction.job.status.CompactionJobCreatedStatus;
 import sleeper.compaction.job.status.CompactionJobStatus;
-import sleeper.compaction.job.status.ProcessRun;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.status.ProcessFinishedStatus;
+import sleeper.core.status.ProcessRun;
 import sleeper.core.status.ProcessStartedStatus;
 
 import java.time.Instant;
