+++ java/compaction-core/src/test/java/sleeper/compaction/job/CompactionJobStatusTaskIdAssignedTest.java
@@ -18,7 +18,7 @@ package sleeper.compaction.job;
 import org.junit.Test;
 import sleeper.compaction.job.status.CompactionJobCreatedStatus;
 import sleeper.compaction.job.status.CompactionJobStatus;
-import sleeper.compaction.job.status.ProcessRun;
+import sleeper.core.status.ProcessRun;
 import sleeper.core.status.ProcessStartedStatus;
 
 import java.time.Instant;
