+++ java/compaction-core/src/test/java/sleeper/compaction/job/CompactionJobStatusesBuilderTest.java
@@ -20,10 +20,10 @@ import sleeper.compaction.job.status.CompactionJobCreatedStatus;
 import sleeper.compaction.job.status.CompactionJobStatus;
 import sleeper.compaction.job.status.CompactionJobStatusUpdateRecord;
 import sleeper.compaction.job.status.CompactionJobStatusesBuilder;
-import sleeper.compaction.job.status.ProcessRun;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.status.ProcessFinishedStatus;
+import sleeper.core.status.ProcessRun;
 import sleeper.core.status.ProcessStartedStatus;
 
 import java.time.Instant;
